adding modifications for augmentation, topology, shearing, run.sh

+++ egs/wsj/s5/utils/copy_data_dir.sh
@@ -86,10 +86,16 @@ fi
 
 if [ -f $srcdir/segments ]; then
   utils/apply_map.pl -f 1 $destdir/utt_map <$srcdir/segments >$destdir/segments
-  cp $srcdir/wav.scp $destdir
+  if [ -f $srcdir/wav.scp ]; then
+    cp $srcdir/wav.scp $destdir
+  elif [ -f $srcdir/images.scp ]; then
+    cp $srcdir/images.scp $destdir
+  fi
 else # no segments->wav indexed by utt.
   if [ -f $srcdir/wav.scp ]; then
     utils/apply_map.pl -f 1 $destdir/utt_map <$srcdir/wav.scp >$destdir/wav.scp
+  elif [ -f $srcdir/images.scp ]; then
+    utils/apply_map.pl -f 1 $destdir/utt_map <$srcdir/images.scp >$destdir/images.scp
   fi
 fi
 
@@ -126,7 +132,7 @@ rm $destdir/spk_map $destdir/utt_map
 
 echo "$0: copied data from $srcdir to $destdir"
 
-for f in feats.scp cmvn.scp vad.scp utt2lang utt2uniq utt2dur utt2num_frames text wav.scp reco2file_and_channel stm glm ctm; do
+for f in feats.scp cmvn.scp vad.scp utt2lang utt2uniq utt2dur utt2num_frames text wav.scp images.scp reco2file_and_channel stm glm ctm; do
   if [ -f $destdir/$f ] && [ ! -f $srcdir/$f ]; then
     echo "$0: file $f exists in dest $destdir but not in src $srcdir.  Moving it to"
     echo " ... $destdir/.backup/$f"
